feat: implement organization dashboard and transcription repositories
- Add OrganizationDashboardResponse DTO for comprehensive dashboard data
- Implement organization dashboard functionality in OrganizationService
- Add transcription repositories (TranscriptionRepository, TranscriptionSessionRepository)
- Update subscription usage tracking and analytics
- Enhance subscription repository with usage tracking methods
- Add comprehensive unit tests for organization dashboard functionality
- Update dependency injection for new repositories
- Improve base multi-tenant repository patterns
- Update task documentation to reflect completed features

This commit adds the foundation for organization dashboard analytics and
transcription management capabilities, building on the standardized controller patterns.

+++ SermonTranscription.Application/DTOs/SubscriptionUsageResponse.cs
@@ -15,7 +15,7 @@ public class SubscriptionUsageResponse
     public int MonthlyLimit { get; set; }
     public int MinutesUsed { get; set; }
     public int MinutesRemaining { get; set; }
-    public decimal TotalUsage { get; set; }
+    public int TotalUsage { get; set; }
     public decimal UsagePercentage { get; set; }
 
     // Usage tracking
